refactor: rename _meta.lua to _options.lua
